Fetching data only if authenticated.
